Move AwsExporterMetrics
This is required to restore the metrics count for RDS requests. Since the describe methods where move to the *All Methods, we can't reliable call the number of counts in the rds.go anymore. So move the exporter to the awsclient package.
